test: 중복적으로 사용하는 예외 발생 모듈 테스트
