+++ __tests__/LottoTest.js
@@ -2,8 +2,6 @@ const { Console } = require('@woowacourse/mission-utils');
 
 const Lotto = require('../src/Lotto');
 
-afterAll(() => Console.close());
-
 describe('로또 클래스 테스트', () => {
   test('로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.', () => {
     expect(() => {
